requirements.txt
socket
time
multiprocessing
sqlite3
python-dotenv
psutil
scapy
struct
